+++ R/notestar-package.R
@@ -4,6 +4,7 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
+#' @importFrom rlang %||%
 #' @importFrom utils capture.output
 ## usethis namespace: end
 NULL